Add detection of Red Hat's hypervisor
Red Hat sets the following fields in SMBIOS information, allowing us
to detect that it's the branded Red Hat hypervisor.

  Manufacturer: Red Hat
  Product Name: KVM
  Version: RHEL 7.6.0 PC (i440FX + PIIX, 1996)

Note that "RHEL 7.6.0" there is the machine type, which is not the
same as the hypervisor version (although the two have a relationship).

Thanks: Timothy Smith at IBM, Karen Noel